tests/wpt/web-platform-tests/fetch/sec-metadata/embed.tentative.https.sub.html

<!DOCTYPE html>
<meta charset="utf-8"/>
<link rel="author" href="mtrzos@google.com" title="Maciek Trzos">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/fetch/sec-metadata/resources/helper.js></script>
<script src=/common/utils.js></script>
<body>
<script>
  let nonce = token();

  promise_test(t => {
    return new Promise((resolve, reject) => {
      let key = "embed-same-origin" + nonce;

      let e = document.createElement('embed');
      e.src = "https://{{host}}:{{ports[https][0]}}/fetch/sec-metadata/resources/record-header.py?file=" + key;
      e.onload = e => {
        let expected = {"dest":"embed", "site":"same-origin", "user":"", "mode":"no-cors"};
        fetch("/fetch/sec-metadata/resources/record-header.py?retrieve=true&file=" + key)
          .then(response => response.text())
          .then(text => assert_header_equals(text, expected))
          .then(_ => resolve())
          .catch(e => reject(e));
      };

      document.body.appendChild(e);
    })
  }, "Same-Origin embed");

  promise_test(t => {
    return new Promise((resolve, reject) => {
      let key = "embed-same-site" + nonce;

      let e = document.createElement('embed');
      e.src = "https://{{hosts[][www]}}:{{ports[https][0]}}/fetch/sec-metadata/resources/record-header.py?file=" + key;
      e.onload = e => {
        let expected = {"dest":"embed", "site":"same-site", "user":"", "mode":"no-cors"};
        fetch("/fetch/sec-metadata/resources/record-header.py?retrieve=true&file=" + key)
          .then(response => response.text())
          .then(text => assert_header_equals(text, expected))
          .then(_ => resolve())
          .catch(e => reject(e));
      };

      document.body.appendChild(e);
    })
  }, "Same-Site embed");

  promise_test(t => {
    return new Promise((resolve, reject) => {
      let key = "embed-cross-site" + nonce;

      let e = document.createElement('embed');
      e.src = "https://{{hosts[alt][www]}}:{{ports[https][0]}}/fetch/sec-metadata/resources/record-header.py?file=" + key;
      e.onload = e => {
        let expected = {"dest":"embed", "site":"cross-site", "user":"", "mode":"no-cors"};
        fetch("/fetch/sec-metadata/resources/record-header.py?retrieve=true&file=" + key)
          .then(response => response.text())
          .then(text => assert_header_equals(text, expected))
          .then(_ => resolve())
          .catch(e => reject(e));
      };

      document.body.appendChild(e);
    })
  }, "Cross-Site embed");
</script>
