third_party/blink/web_tests/external/wpt/fetch/sec-metadata/embed.tentative.https.sub-expected.txt

This is a testharness.js-based test.
FAIL Same-Origin embed assert_equals: dest expected "embed" but got "nested-document"
FAIL Same-Site embed assert_equals: dest expected "embed" but got "nested-document"
FAIL Cross-Site embed assert_equals: dest expected "embed" but got "nested-document"
Harness: the test ran to completion.
